fix: #67-expand-norm
